Fixed 2 problems for GitSessionSave
If you are working on a repo with a lot of files (1000+) the old method
of checking if we are on a git repo using git status is super slow.
I've changed the function in_git_repo to make it use
git rev-parse --git-dir because it is much faster.
Also added a call to in_git_repo before attempting to save a new
session.